requirements.txt
fastapi
uvicorn
httpx
openai
python-dotenv
faiss-cpu==1.7.4
numpy==1.23.5
sentence-transformers==2.2.0
torch==2.1.2
nltk>=3.8
scipy>=1.10
huggingface_hub>=0.20.1